+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 Started Change log file as per release 0.0.0.54
 ## AGMPowerLIB  (0.0.0.59)
 * Added Restore-AGMLibSAPHANA
-* Minor fixes to Mount-AGMLibSAPHANA
+* Minor fixes to Mount-AGMLibSAPHANA including CSV output and add  New-AGMLibSAPHANAMultiMount
+* Minor fixes to New-AGMLibGCPInstance 
 
 ## AGMPowerLIB  (0.0.0.58)
 * Get-AGMLibPolicies - added options